refactor: 예약 삭제 시 상태코드 204 반환하도록 변경
